chore: extract Dice subclasses to separate files
