+++ src/app/Dice.ts
@@ -8,41 +8,3 @@ export class Dice implements GetRandomValue {
     return sample(this.diceSides) as AnimalNames;
   }
 }
-
-export class FirstDice extends Dice {
-  constructor() {
-    super([
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.SHEEP,
-      AnimalNames.SHEEP,
-      AnimalNames.PIG,
-      AnimalNames.PIG,
-      AnimalNames.HORSE,
-      AnimalNames.FOX,
-    ]);
-  }
-}
-
-export class SecondDice extends Dice {
-  constructor() {
-    super([
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.RABBIT,
-      AnimalNames.SHEEP,
-      AnimalNames.SHEEP,
-      AnimalNames.SHEEP,
-      AnimalNames.PIG,
-      AnimalNames.COW,
-      AnimalNames.WOLF,
-    ]);
-  }
-}
